+++ core/src/main/kotlin/de/sambalmueslie/gschwind/core/api/Stream.kt
@@ -2,7 +2,10 @@ package de.sambalmueslie.gschwind.core.api
 
 interface Stream {
     val id: String
+
     fun print()
+
     fun start()
     fun stop()
+
 }
