+++ core/src/main/kotlin/de/sambalmueslie/gschwind/core/builder/OperatorStreamBuilder.kt
@@ -1,8 +1,8 @@
 package de.sambalmueslie.gschwind.core.builder
 
-import de.sambalmueslie.gschwind.core.job.EmitterWrapper
-import de.sambalmueslie.gschwind.core.job.OperatorWrapper
-import de.sambalmueslie.gschwind.core.job.StreamWrapper
+import de.sambalmueslie.gschwind.core.wrapper.EmitterWrapper
+import de.sambalmueslie.gschwind.core.wrapper.OperatorWrapper
+import de.sambalmueslie.gschwind.core.wrapper.StreamWrapper
 
 internal class OperatorStreamBuilder<R, E>(
     stream: StreamWrapper,
