+++ core/src/main/kotlin/de/sambalmueslie/gschwind/core/builder/SourceStreamBuilder.kt
@@ -1,9 +1,9 @@
 package de.sambalmueslie.gschwind.core.builder
 
 import de.sambalmueslie.gschwind.core.api.Source
-import de.sambalmueslie.gschwind.core.job.EmitterWrapper
-import de.sambalmueslie.gschwind.core.job.SourceWrapper
-import de.sambalmueslie.gschwind.core.job.StreamWrapper
+import de.sambalmueslie.gschwind.core.wrapper.EmitterWrapper
+import de.sambalmueslie.gschwind.core.wrapper.SourceWrapper
+import de.sambalmueslie.gschwind.core.wrapper.StreamWrapper
 
 internal class SourceStreamBuilder<T>(
     stream: StreamWrapper,
