--- core/src/main/kotlin/de/sambalmueslie/gschwind/core/builder/StreamDsl.kt
@@ -1,4 +1,0 @@
-package de.sambalmueslie.gschwind.core.builder
-
-@DslMarker
-annotation class StreamDsl
