+++ core/src/main/kotlin/de/sambalmueslie/gschwind/core/wrapper/SinkWrapper.kt
@@ -1,4 +1,4 @@
-package de.sambalmueslie.gschwind.core.job
+package de.sambalmueslie.gschwind.core.wrapper
 
 import de.sambalmueslie.gschwind.core.api.Sink
 import de.sambalmueslie.gschwind.core.api.StreamElement
